Reorganize Date extensions and arrangers
