+++ RecentActivity/Tests/RecentActivityTests/Specs/Model/MonthGroupKeyTests.swift
@@ -1,6 +1,6 @@
 import Foundation
 import Testing
-import GOVKitTestUtilities
+@testable import GOVKitTestUtilities
 
 @testable import RecentActivity
